Put mtheme in a zip file, removing resources dir
I ran into a bug where minted wouldn't respect the -output-directory
parameter and would create files inside the resources dir. The only
solution seemed to be configuring minted to use the output directory by
adding a line in the source tex file. Since I'm generating tex source
and compiling pdfs in separate steps, this is not good.

And I couldn't let Xelatex run in the resources dir because it produces
too much junk, and manually removing it is risky.

So I decided to pack everything from the resources dir (which was
actually only the beamer mtheme) into a zip, which is extracted in the
temporary directory created during compilation. This works well. It's
wasteful, but latex compilation is so slow you don't even notice.